+++ router/usecase/router_usecase_test.go
@@ -32,6 +32,7 @@ const (
 	poolIDOneBalancer      = uint64(1)
 	poolID1135Concentrated = uint64(1135)
 	poolID1265Concentrated = uint64(1265)
+	poolID1399Concentrated = uint64(1399)
 	poolID1400Concentrated = uint64(1400)
 )
 
@@ -480,10 +481,11 @@ func (s *RouterTestSuite) TestConvertRankedToCandidateRoutes() {
 // We restrict the number of routes via config.
 //
 // As of today there are 3 major ATOM / OSMO pools:
-// Pool ID 1: https://app.osmosis.zone/pool/1 (balancer) 0.2% spread factor and 18M of liquidity to date
-// Pool ID 1135: https://app.osmosis.zone/pool/1135 (concentrated) 0.2% spread factor and 10M of liquidity to date
-// Pool ID 1265: https://app.osmosis.zone/pool/1265 (concentrated) 0.05% spread factor and 1.4M of liquidity to date
-// Pool ID 1400: https://app.osmosis.zone/pool/1400 (concentrated) 0.00% spread factor and 322K of liquidity to date
+// Pool ID 1: https://app.osmosis.zone/pool/1 (balancer) 0.2% spread factor and 11M of liquidity to date
+// Pool ID 1135: https://app.osmosis.zone/pool/1135 (concentrated) 0.2% spread factor and 6.34M of liquidity to date
+// Pool ID 1265: https://app.osmosis.zone/pool/1265 (concentrated) 0.05% spread factor and 435K of liquidity to date
+// Pool ID 1399: https://app.osmosis.zone/pool/1399 (concentrated) 0.01% spread factor and 78.5K of liquidity to date
+// Pool ID 1400: https://app.osmosis.zone/pool/1400 (concentrated) 0.00% spread factor and 384K of liquidity to date
 //
 // Based on this state, the small amounts of token in should go through pool 1265
 // Medium amounts of token in should go through pool 1135
@@ -509,9 +511,9 @@ func (s *RouterTestSuite) TestGetOptimalQuote_Cache_Overwrites() {
 		"cache is not set, computes routes": {
 			amountIn: defaultAmountInCache,
 
-			// For the default amount in, we expect pool 1400 to be returned.
+			// For the default amount in, we expect pool 1399 to be returned.
 			// See test description above for details.
-			expectedRoutePoolID: poolID1400Concentrated,
+			expectedRoutePoolID: poolID1399Concentrated,
 		},
 		"cache is set to balancer - overwrites computed": {
 			amountIn: defaultAmountInCache,
@@ -533,7 +535,7 @@ func (s *RouterTestSuite) TestGetOptimalQuote_Cache_Overwrites() {
 			cacheExpiryDuration: time.Nanosecond,
 
 			// We expect pool 1400 because the cache with balancer pool expires.
-			expectedRoutePoolID: poolID1400Concentrated,
+			expectedRoutePoolID: poolID1399Concentrated,
 		},
 	}
 
@@ -551,7 +553,7 @@ func (s *RouterTestSuite) TestGetOptimalQuote_Cache_Overwrites() {
 			}
 
 			// Mock router use case.
-			mainnetUseCase := s.SetupRouterAndPoolsUsecase(router, mainnetState, rankedRouteCache, candidateRouteCache)
+			mainnetUseCase := s.SetupRouterAndPoolsUsecase(router, mainnetState, routertesting.WithRankedRoutesCache(rankedRouteCache), routertesting.WithCandidateRoutesCache(candidateRouteCache))
 
 			// System under test
 			quote, err := mainnetUseCase.Router.GetOptimalQuote(context.Background(), sdk.NewCoin(defaultTokenInDenom, tc.amountIn), defaultTokenOutDenom)
@@ -601,7 +603,7 @@ func (s *RouterTestSuite) TestGetCandidateRoutes_Chain_FindUnsupportedRoutes() {
 
 	// Set up mainnet mock state.
 	router, mainnetState := s.SetupMainnetRouter(*config.Router, *config.Pricing)
-	mainnetUsecase := s.SetupRouterAndPoolsUsecase(router, mainnetState, cache.New(), cache.New())
+	mainnetUsecase := s.SetupRouterAndPoolsUsecase(router, mainnetState)
 
 	tokenMetadata, err := mainnetUsecase.Tokens.GetFullTokenMetadata(context.Background())
 	s.Require().NoError(err)
@@ -639,7 +641,7 @@ func (s *RouterTestSuite) TestGetCandidateRoutes_Chain_FindUnsupportedRoutes() {
 	config.Router.MinOSMOLiquidity = 0
 	// Set up mainnet mock state.
 	router, mainnetState = s.SetupMainnetRouter(*config.Router, *config.Pricing)
-	mainnetUsecase = s.SetupRouterAndPoolsUsecase(router, mainnetState, cache.New(), cache.New())
+	mainnetUsecase = s.SetupRouterAndPoolsUsecase(router, mainnetState)
 	for chainDenom, tokenMeta := range tokenMetadata {
 
 		routes, err := router.GetCandidateRoutes(chainDenom, USDC)
@@ -679,7 +681,7 @@ func (s *RouterTestSuite) TestPriceImpactRoute_Fractions() {
 
 	// Set up mainnet mock state.
 	router, mainnetState := s.SetupMainnetRouter(*config.Router, *config.Pricing)
-	mainnetUsecase := s.SetupRouterAndPoolsUsecase(router, mainnetState, cache.New(), cache.New())
+	mainnetUsecase := s.SetupRouterAndPoolsUsecase(router, mainnetState)
 
 	tokenMetadata, err := mainnetUsecase.Tokens.GetFullTokenMetadata(context.Background())
 
@@ -717,7 +719,7 @@ func (s *RouterTestSuite) TestSortPools() {
 
 	router, mainnetState := s.SetupDefaultMainnetRouter()
 
-	usecase := s.SetupRouterAndPoolsUsecase(router, mainnetState, cache.New(), cache.New())
+	usecase := s.SetupRouterAndPoolsUsecase(router, mainnetState)
 
 	pools, err := usecase.Pools.GetAllPools()
 	s.Require().NoError(err)